Refactor files to be in correctly nested packages (#6120)
