+++ pmd-exclude.properties
@@ -55,7 +55,6 @@ com.thealgorithms.maths.SumOfArithmeticSeries=UselessParentheses
 com.thealgorithms.maths.TrinomialTriangle=UselessParentheses
 com.thealgorithms.maths.VampireNumber=CollapsibleIfStatements
 com.thealgorithms.maths.Volume=UselessParentheses
-com.thealgorithms.matrixexponentiation.Fibonacci=UnnecessaryFullyQualifiedName
 com.thealgorithms.misc.Sparsity=UselessParentheses
 com.thealgorithms.misc.ThreeSumProblem=UselessParentheses
 com.thealgorithms.misc.WordBoggle=UselessParentheses
