+++ src/main/java/com/thealgorithms/matrix/MatrixTranspose.java
@@ -1,4 +1,4 @@
-package com.thealgorithms.misc;
+package com.thealgorithms.matrix;
 
 /**
  *
