+++ src/main/java/com/thealgorithms/matrix/MirrorOfMatrix.java
@@ -1,4 +1,4 @@
-package com.thealgorithms.misc;
+package com.thealgorithms.matrix;
 
 // Problem Statement
 /*
