+++ src/test/java/com/thealgorithms/matrix/InverseOfMatrixTest.java
@@ -1,5 +1,4 @@
-package com.thealgorithms.misc;
-
+package com.thealgorithms.matrix;
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 
 import java.util.stream.Stream;
