+++ src/test/java/com/thealgorithms/matrix/MedianOfMatrixTest.java
@@ -1,4 +1,4 @@
-package com.thealgorithms.misc;
+package com.thealgorithms.matrix;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
